Struct near_jsonrpc_client::methods::EXPERIMENTAL_protocol_config::ProtocolConfigView

pub struct ProtocolConfigView {

    pub protocol_version: u32,
    pub genesis_time: DateTime<Utc>,
    pub chain_id: String,
    pub genesis_height: u64,
    pub num_block_producer_seats: u64,
    pub num_block_producer_seats_per_shard: Vec<u64, Global>,
    pub avg_hidden_validator_seats_per_shard: Vec<u64, Global>,
    pub dynamic_resharding: bool,
    pub protocol_upgrade_stake_threshold: Ratio<isize>,
    pub epoch_length: u64,
    pub gas_limit: u64,
    pub min_gas_price: u128,
    pub max_gas_price: u128,
    pub block_producer_kickout_threshold: u8,
    pub chunk_producer_kickout_threshold: u8,
    pub online_min_threshold: Ratio<isize>,
    pub online_max_threshold: Ratio<isize>,
    pub gas_price_adjustment_rate: Ratio<isize>,
    pub runtime_config: RuntimeConfig,
    pub transaction_validity_period: u64,
    pub protocol_reward_rate: Ratio<isize>,
    pub max_inflation_rate: Ratio<isize>,
    pub num_blocks_per_year: u64,
    pub protocol_treasury_account: AccountId,
    pub fishermen_threshold: u128,
    pub minimum_stake_divisor: u64,

}

Fields
protocol_version: u32

Current Protocol Version

genesis_time: DateTime<Utc>

Official time of blockchain start.

chain_id: String

ID of the blockchain. This must be unique for every blockchain. If your testnet blockchains do not have unique chain IDs, you will have a bad time.

genesis_height: u64

Height of genesis block.

num_block_producer_seats: u64

Number of block producer seats at genesis.

num_block_producer_seats_per_shard: Vec<u64, Global>

Defines number of shards and number of block producer seats per each shard at genesis.

avg_hidden_validator_seats_per_shard: Vec<u64, Global>

Expected number of hidden validators per shard.

dynamic_resharding: bool

Enable dynamic re-sharding.

protocol_upgrade_stake_threshold: Ratio<isize>

Threshold of stake that needs to indicate that they ready for upgrade.

epoch_length: u64

Epoch length counted in block heights.

gas_limit: u64

Initial gas limit.

min_gas_price: u128

Minimum gas price. It is also the initial gas price.

max_gas_price: u128

Maximum gas price.

block_producer_kickout_threshold: u8

Criterion for kicking out block producers (this is a number between 0 and 100)

chunk_producer_kickout_threshold: u8

Criterion for kicking out chunk producers (this is a number between 0 and 100)

online_min_threshold: Ratio<isize>

Online minimum threshold below which validator doesn’t receive reward.

online_max_threshold: Ratio<isize>

Online maximum threshold above which validator gets full reward.

gas_price_adjustment_rate: Ratio<isize>

Gas price adjustment rate

runtime_config: RuntimeConfig

Runtime configuration (mostly economics constants).

transaction_validity_period: u64

Number of blocks for which a given transaction is valid

protocol_reward_rate: Ratio<isize>

Protocol treasury rate

max_inflation_rate: Ratio<isize>

Maximum inflation on the total supply every epoch.

num_blocks_per_year: u64

Expected number of blocks per year

protocol_treasury_account: AccountId

Protocol treasury account

fishermen_threshold: u128

Fishermen stake threshold.

minimum_stake_divisor: u64

The minimum stake required for staking is last seat price divided by this number.